chore: separate 30 and 31 deployer in helm
